smartmed: Separate NAND from smartmed
