+++ src/mame/misc/crospuzl.cpp
@@ -370,8 +370,7 @@ void crospuzl_state::crospuzl(machine_config &config)
 
 //  ROM strings have references to a K9FXX08 device
 //  TODO: use this device, in machine/smartmed.h (has issues with is_busy() emulation)
-//  NAND(config, m_nand, 0);
-//  m_nand->set_nand_type(nand_device::chip::K9F1G08U0B); // TODO: exact flavor
+//  SAMSUNG_K9F1G08U0B(config, m_nand, 0); // TODO: exact flavor
 
 	PCF8583(config, m_rtc, 32.768_kHz_XTAL);
 }
